- add input arguments and update by date option for metadata_dump_parser

+++ data_extraction/caselaw/rechtspraak/rechtspraak_metadata_dump_parser.py
@@ -1,4 +1,4 @@
-from os.path import dirname, abspath
+from os.path import dirname, abspath, basename, exists
 import sys
 sys.path.append(dirname(dirname(dirname(dirname(abspath(__file__))))))
 from lxml import etree
@@ -10,15 +10,43 @@ import time
 from definitions.storage_handler import Storage, DIR_RECHTSPRAAK, CSV_RS_CASES, CSV_RS_OPINIONS, CSV_RS_ECLIS
 from definitions.terminology.field_names import SOURCE, JURISDICTION_COUNTRY, ECLI_DECISION
 from definitions.terminology.field_values import RECHTSPRAAK, NL
+import argparse
+import datetime
+
+parser = argparse.ArgumentParser()
+parser.add_argument('storage', choices=['local', 'aws'], help='location to take input data from and save output data to')
+parser.add_argument('-d', '--date', type=datetime.date.fromisoformat, help='start decision date to fetch data from')
+parser.add_argument('-u', '--update', action='store_true', help='update data from most recent decision date on storage')
+args = parser.parse_args()
+storage = Storage(args.storage)
+print('Input/Output data storage:', args.storage)
+
+input_path = DIR_RECHTSPRAAK
+output_path_cases = CSV_RS_CASES
+output_path_opinions = CSV_RS_OPINIONS
+output_path_eclis = CSV_RS_ECLIS
+
+if args.update:
+    from_date_cases = storage.last_updated(output_path_cases, 'date')
+    from_date_opinions = storage.last_updated(output_path_opinions, 'date')
+    from_date_eclis = storage.last_updated(output_path_eclis, 'date')
+    print(from_date_cases)
+    print(from_date_opinions)
+    print(from_date_eclis)
+    from_date = min(from_date_cases, from_date_opinions, from_date_opinions)
+elif args.date:
+    from_date = args.date
+else:
+    from_date = datetime.date(1900, 1, 1)
+
+print('Data will be processed from:', from_date.isoformat())
 
 is_case = False
 
-case_records = []
-opinion_records = []
+case_counter = 0
+opinion_counter = 0
+datarecord = dict()
 
-global case_counter
-global opinion_counter
-global datarecord
 
 # Field names used for output csv. Field names correspond to tags of original data
 IDENTIFIER = 'identifier'
@@ -165,14 +193,20 @@ def processtag(cleantagname, tag):
 # write column names to csv
 def initialise_csv_files():
     initialise_data_record()
-    with open(CSV_RS_OPINIONS, 'w') as f:
-        writer = csv.DictWriter(f, datarecord.keys())
-        writer.writeheader()
-    with open(CSV_RS_CASES, 'w') as f:
-        # Using dictionary keys as fieldnames for the CSV file header
-        datarecord.pop(ECLI_DECISION)
-        writer = csv.DictWriter(f, datarecord.keys())
-        writer.writeheader()
+    if not exists(CSV_RS_OPINIONS):
+        with open(CSV_RS_OPINIONS, 'w') as f:
+            writer = csv.DictWriter(f, datarecord.keys())
+            writer.writeheader()
+    if not exists(CSV_RS_CASES):
+        with open(CSV_RS_CASES, 'w') as f:
+            # Using dictionary keys as fieldnames for the CSV file header
+            datarecord.pop(ECLI_DECISION)
+            writer = csv.DictWriter(f, datarecord.keys())
+            writer.writeheader()
+    if not exists(CSV_RS_ECLIS):
+        with open(CSV_RS_ECLIS, 'w') as f:
+            writer = csv.DictWriter(f, ['ecli', 'date'])
+            writer.writeheader()
 
 
 # Function to write new line in the csv file
@@ -212,8 +246,7 @@ def parse_metadata_from_xml_file(filename):
         datarecord.pop(ECLI_DECISION)
         write_line_csv(CSV_RS_CASES, datarecord)
         # write case eclis to separate .csv file to later be able to run case citation extractor script
-        with open(CSV_RS_ECLIS, 'a') as f:
-            f.write(datarecord[IDENTIFIER] + '\n')
+        write_line_csv(CSV_RS_ECLIS, {'ecli': datarecord[IDENTIFIER], 'date': datarecord[DATE]})
     else:
         opinion_counter += 1
         print("\033[95mOPINION\033[0m %s" % datarecord[IDENTIFIER])
@@ -222,21 +255,29 @@ def parse_metadata_from_xml_file(filename):
 # Start script timer
 start = time.time()
 
+print(f"Fetching data from {args.storage} storage...")
+storage.fetch(input_path)
+
 print("Building index of XML files...\n")
 
 list_of_files_to_parse = []
 
 # List all top-level directories
-file_tree = os.walk(DIR_RECHTSPRAAK)
+file_tree = os.walk(input_path)
 
 # List all the files from the sub-directories
 for (dirpath, dirnames, filenames) in file_tree:
-    for file in filenames:
-        # Append only the xml files
-        if file[-4:].lower() == ".xml":
-            list_of_files_to_parse.append(os.path.join(dirpath, file))
-        else:
-            print(file)
+    if filenames:
+        year, month = int(basename(dirpath)[:4]), int(basename(dirpath)[4:])
+        # skip files before from date
+        if year < from_date.year or year == from_date.year and month < from_date.month:
+            continue
+        for file in filenames:
+            # Append only the xml files
+            if file[-4:].lower() == ".xml":
+                list_of_files_to_parse.append(os.path.join(dirpath, file))
+            else:
+                print(file)
 
     # list_of_files_to_parse.extend([os.path.join(dirpath, file) for file in filenames])
 
@@ -255,11 +296,16 @@ for file in list_of_files_to_parse:
     parse_metadata_from_xml_file(file)
     index += 1
 
-print("Number of files: %i %i+%i" % (num_files, len(case_records), len(opinion_records)))
+print("Number of files: %i %i+%i" % (num_files, case_counter, opinion_counter))
 
 print()
 print("Parsing successfully completed!")
 
+
+print(f"Updating {args.storage} storage ...")
+storage.update(output_path_cases)
+storage.update(output_path_opinions)
+storage.update(output_path_eclis)
 # Stop the script timer
 end = time.time()
